requirements.txt
Django==5.0.6
djangorestframework==3.15.1
psycopg==3.1.19
drf-yasg==1.21.7
rest-framework-simplejwt==0.0.2

